修复由uint_fast32_t导致的跨平台错误
详细见:https://www.luogu.com.cn/discuss/1105383

+++ vinteger_bit_operation.cpp
@@ -71,7 +71,7 @@ namespace algae
             int stop = 0;
             while(__buffer[stop] == 0)
             {   
-                if((++stop) >= length)
+                if((++stop) >= (int)length)
                     break;
             }
